backend/requirements.txt
Django==4.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-allauth==0.61.1
dj-rest-auth==5.0.2
Pillow==10.2.0
boto3==1.34.34
azure-mgmt-desktopvirtualization==1.0.0
azure-identity==1.15.0
django-cryptography==1.1
python-dotenv==1.0.0
gunicorn==21.2.0

